backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
sqlalchemy
numpy
scikit-learn
joblib
pinecone
psycopg[binary]
transformers
torch
